refactor codelens features to smaller pieces
