+++ frontend/src/components/MonacoDiffEditor/index.ts
@@ -1,0 +1,3 @@
+import MonacoDiffEditor from './MonacoDiffEditor'
+
+export default MonacoDiffEditor
